merge

+++ src/render/src/shaders/hips/raytracer/raytracer.vert
@@ -1,6 +1,6 @@
 #version 300 es
-precision mediump float;
-precision mediump int;
+precision highp float;
+precision highp int;
 
 layout (location = 0) in vec2 pos_clip_space;
 layout (location = 1) in vec2 lonlat;
@@ -12,7 +12,6 @@ out vec2 out_lonlat;
 uniform mat4 model;
 uniform vec2 ndc_to_clip;
 uniform float czf;
-
 
 void main() {
     gl_Position = vec4(pos_clip_space / (ndc_to_clip * czf), 0.0, 1.0);
